♻️ (server): Extract common logic to common service
