+++ apps/server/src/bookmarks/usecases/add-bookmark/add-bookmark.usecase.ts
@@ -2,34 +2,32 @@ import { InjectRepository } from '@nestjs/typeorm';
 import { Usecase } from '@readable/common/usecase';
 import { Repository } from 'typeorm';
 import { AddBookMarkInput } from './add-bookmark.input';
-import * as ogs from 'open-graph-scraper';
-import { BookmarkBuilder } from '@readable/bookmarks/infrastructures/typeorm/bookmark.entity.builder';
 import { Bookmark as BookmarkEntity } from '@readable/bookmarks/infrastructures/typeorm/bookmark.entity';
 import { Bookmark as BookmarkModel } from '@readable/bookmarks/models/bookmark.model';
+import { BookmarksService } from '@readable/bookmarks/bookmarks.service';
+import * as sha256 from 'crypto-js/sha256';
 
 export class AddBookmarkUsecase implements Usecase<AddBookMarkInput, BookmarkModel> {
-  constructor(@InjectRepository(BookmarkEntity) private readonly bookmarksRepository: Repository<BookmarkEntity>) {}
+  constructor(
+    private readonly bookmarksService: BookmarksService,
+    @InjectRepository(BookmarkEntity) private readonly bookmarksRepository: Repository<BookmarkEntity>
+  ) {}
 
   async execute(command: AddBookMarkInput) {
     const { url } = command;
 
-    const ogsOptions = { url };
-    const { result } = await ogs(ogsOptions);
+    const urlHash = sha256(url).toString();
 
-    const bookmark = new BookmarkBuilder()
-      .setUrl(result['ogUrl'] ?? url)
-      .setSiteName(result['ogSiteName'] ?? '')
-      .setTitle(result['ogTitle'] ?? '')
-      .setType(result['ogType'] ?? '')
-      .setImageUrl(result['ogImage']['url'] ?? '')
-      .setDescription(result['ogDescription'] ?? '')
-      .setTags(result['ogTags'] ?? '')
-      .build();
+    const existingBookmark = await this.bookmarksService.getBookmarkByUrlHash(urlHash);
+    if (existingBookmark) return existingBookmark;
 
-    bookmark.urlHash = '1234';
-    bookmark.generatedImage = '...';
+    const bookmarkInfo = await this.bookmarksService.generateBasicBookmarkInfo(command);
+    bookmarkInfo.urlHash = urlHash;
 
-    const newBookmark = this.bookmarksRepository.create(bookmark);
+    // TODO(Teddy): WIP
+    // bookmarkInfo.generatedImage = '...';
+
+    const newBookmark = this.bookmarksRepository.create(bookmarkInfo);
     return this.bookmarksRepository.save(newBookmark);
   }
 }
